update Header to conditionally render searchbar/filters and add to subscriptions
